memoize frame ids, remove old menu, rename new menu, fix menu input left padding
