+++ packages/terra-form-select/src/search/Menu.module.scss
@@ -104,6 +104,7 @@
     background: transparent;
     border: 0;
     color: inherit;
+    flex: 1 1 auto;
     font-family: inherit;
     font-size: var(--terra-form-select-search-focus-font-size, 1.143rem);
     height: var(--terra-form-select-search-focus-height, 1.375em);
